build: moving the logical DNS cluter to extensions
