+++ source/common/upstream/BUILD
@@ -331,40 +331,6 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "logical_host_lib",
-    srcs = ["logical_host.cc"],
-    hdrs = ["logical_host.h"],
-    deps = [
-        ":upstream_includes",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/endpoint/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_library(
-    name = "logical_dns_cluster_lib",
-    srcs = ["logical_dns_cluster.cc"],
-    hdrs = ["logical_dns_cluster.h"],
-    deps = [
-        ":cluster_factory_lib",
-        ":logical_host_lib",
-        ":upstream_includes",
-        "//envoy/upstream:cluster_factory_interface",
-        "//source/common/common:dns_utils_lib",
-        "//source/common/common:empty_string",
-        "//source/common/config:utility_lib",
-        "//source/common/network:address_lib",
-        "//source/common/network:utility_lib",
-        "//source/common/protobuf",
-        "//source/common/protobuf:utility_lib",
-        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/endpoint/v3:pkg_cc_proto",
-    ],
-    alwayslink = LEGACY_ALWAYSLINK,
-)
-
-envoy_cc_library(
     name = "outlier_detection_lib",
     srcs = ["outlier_detection_impl.cc"],
     hdrs = ["outlier_detection_impl.h"],
@@ -519,7 +485,6 @@ envoy_cc_library(
         ":eds_lib",
         ":health_checker_lib",
         # TODO(mattklein123): Move the clusters to extensions so they can be compiled out.
-        ":logical_dns_cluster_lib",
         ":static_cluster_lib",
         ":strict_dns_cluster_lib",
         ":upstream_includes",
