+++ source/extensions/clusters/logical_dns/logical_host.cc
@@ -1,4 +1,4 @@
-#include "source/common/upstream/logical_host.h"
+#include "source/extensions/clusters/logical_dns/logical_host.h"
 
 namespace Envoy {
 namespace Upstream {
