+++ test/common/upstream/BUILD
@@ -444,33 +444,6 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
-    name = "logical_dns_cluster_test",
-    srcs = ["logical_dns_cluster_test.cc"],
-    deps = [
-        ":utility_lib",
-        "//source/common/event:dispatcher_lib",
-        "//source/common/network:utility_lib",
-        "//source/common/upstream:logical_dns_cluster_lib",
-        "//source/common/upstream:upstream_lib",
-        "//source/extensions/transport_sockets/raw_buffer:config",
-        "//source/server:transport_socket_config_lib",
-        "//test/mocks:common_lib",
-        "//test/mocks/local_info:local_info_mocks",
-        "//test/mocks/network:network_mocks",
-        "//test/mocks/protobuf:protobuf_mocks",
-        "//test/mocks/runtime:runtime_mocks",
-        "//test/mocks/server:admin_mocks",
-        "//test/mocks/server:instance_mocks",
-        "//test/mocks/ssl:ssl_mocks",
-        "//test/mocks/thread_local:thread_local_mocks",
-        "//test/mocks/upstream:cluster_manager_mocks",
-        "//test/test_common:utility_lib",
-        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_test(
     name = "outlier_detection_impl_test",
     srcs = ["outlier_detection_impl_test.cc"],
     external_deps = ["abseil_optional"],
